+++ jni/lib/src/lang/jfloat.dart
@@ -9,7 +9,7 @@ import '../third_party/generated_bindings.dart';
 import '../types.dart';
 import 'jnumber.dart';
 
-class JFloatType extends JObjType<JFloat> {
+final class JFloatType extends JObjType<JFloat> {
   const JFloatType();
 
   @override