+++ jnigen/example/in_app_java/lib/android_utils.dart
@@ -54,7 +54,7 @@ class AndroidUtils extends jni.JObject {
   }
 }
 
-class $AndroidUtilsType extends jni.JObjType<AndroidUtils> {
+final class $AndroidUtilsType extends jni.JObjType<AndroidUtils> {
   const $AndroidUtilsType();
 
   @override
@@ -1026,7 +1026,7 @@ class EmojiCompat extends jni.JObject {
   }
 }
 
-class $EmojiCompatType extends jni.JObjType<EmojiCompat> {
+final class $EmojiCompatType extends jni.JObjType<EmojiCompat> {
   const $EmojiCompatType();
 
   @override
@@ -1360,7 +1360,7 @@ class EmojiCompat_Config extends jni.JObject {
   }
 }
 
-class $EmojiCompat_ConfigType extends jni.JObjType<EmojiCompat_Config> {
+final class $EmojiCompat_ConfigType extends jni.JObjType<EmojiCompat_Config> {
   const $EmojiCompat_ConfigType();
 
   @override
@@ -1449,7 +1449,7 @@ class EmojiCompat_MetadataRepoLoaderCallback extends jni.JObject {
   }
 }
 
-class $EmojiCompat_MetadataRepoLoaderCallbackType
+final class $EmojiCompat_MetadataRepoLoaderCallbackType
     extends jni.JObjType<EmojiCompat_MetadataRepoLoaderCallback> {
   const $EmojiCompat_MetadataRepoLoaderCallbackType();
 
@@ -1550,7 +1550,7 @@ class EmojiCompat_GlyphChecker extends jni.JObject {
   }
 }
 
-class $EmojiCompat_GlyphCheckerType
+final class $EmojiCompat_GlyphCheckerType
     extends jni.JObjType<EmojiCompat_GlyphChecker> {
   const $EmojiCompat_GlyphCheckerType();
 
@@ -1612,7 +1612,7 @@ class EmojiCompat_MetadataRepoLoader extends jni.JObject {
   }
 }
 
-class $EmojiCompat_MetadataRepoLoaderType
+final class $EmojiCompat_MetadataRepoLoaderType
     extends jni.JObjType<EmojiCompat_MetadataRepoLoader> {
   const $EmojiCompat_MetadataRepoLoaderType();
 
@@ -1696,7 +1696,7 @@ class EmojiCompat_InitCallback extends jni.JObject {
   }
 }
 
-class $EmojiCompat_InitCallbackType
+final class $EmojiCompat_InitCallbackType
     extends jni.JObjType<EmojiCompat_InitCallback> {
   const $EmojiCompat_InitCallbackType();
 
@@ -1770,7 +1770,7 @@ class EmojiCompat_DefaultSpanFactory extends jni.JObject {
   }
 }
 
-class $EmojiCompat_DefaultSpanFactoryType
+final class $EmojiCompat_DefaultSpanFactoryType
     extends jni.JObjType<EmojiCompat_DefaultSpanFactory> {
   const $EmojiCompat_DefaultSpanFactoryType();
 
@@ -1840,7 +1840,7 @@ class EmojiCompat_SpanFactory extends jni.JObject {
   }
 }
 
-class $EmojiCompat_SpanFactoryType
+final class $EmojiCompat_SpanFactoryType
     extends jni.JObjType<EmojiCompat_SpanFactory> {
   const $EmojiCompat_SpanFactoryType();
 
@@ -1938,7 +1938,7 @@ class DefaultEmojiCompatConfig extends jni.JObject {
   }
 }
 
-class $DefaultEmojiCompatConfigType
+final class $DefaultEmojiCompatConfigType
     extends jni.JObjType<DefaultEmojiCompatConfig> {
   const $DefaultEmojiCompatConfigType();
 
@@ -2016,7 +2016,7 @@ class DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API28
   }
 }
 
-class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API28Type
+final class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API28Type
     extends jni
     .JObjType<DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API28> {
   const $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API28Type();
@@ -2125,7 +2125,7 @@ class DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API19
   }
 }
 
-class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API19Type
+final class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API19Type
     extends jni
     .JObjType<DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API19> {
   const $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper_API19Type();
@@ -2263,7 +2263,8 @@ class DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper
   }
 }
 
-class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelperType extends jni
+final class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelperType
+    extends jni
     .JObjType<DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelper> {
   const $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigHelperType();
 
@@ -2352,7 +2353,8 @@ class DefaultEmojiCompatConfig_DefaultEmojiCompatConfigFactory
   }
 }
 
-class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigFactoryType extends jni
+final class $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigFactoryType
+    extends jni
     .JObjType<DefaultEmojiCompatConfig_DefaultEmojiCompatConfigFactory> {
   const $DefaultEmojiCompatConfig_DefaultEmojiCompatConfigFactoryType();
 
@@ -2460,7 +2462,7 @@ class Build_Partition extends jni.JObject {
   }
 }
 
-class $Build_PartitionType extends jni.JObjType<Build_Partition> {
+final class $Build_PartitionType extends jni.JObjType<Build_Partition> {
   const $Build_PartitionType();
 
   @override
@@ -2612,7 +2614,7 @@ class Build_VERSION extends jni.JObject {
   }
 }
 
-class $Build_VERSIONType extends jni.JObjType<Build_VERSION> {
+final class $Build_VERSIONType extends jni.JObjType<Build_VERSION> {
   const $Build_VERSIONType();
 
   @override
@@ -2762,7 +2764,7 @@ class Build_VERSION_CODES extends jni.JObject {
   }
 }
 
-class $Build_VERSION_CODESType extends jni.JObjType<Build_VERSION_CODES> {
+final class $Build_VERSION_CODESType extends jni.JObjType<Build_VERSION_CODES> {
   const $Build_VERSION_CODESType();
 
   @override
@@ -3111,7 +3113,7 @@ class Build extends jni.JObject {
   }
 }
 
-class $BuildType extends jni.JObjType<Build> {
+final class $BuildType extends jni.JObjType<Build> {
   const $BuildType();
 
   @override
@@ -3623,7 +3625,7 @@ class HashMap<$K extends jni.JObject, $V extends jni.JObject>
   }
 }
 
-class $HashMapType<$K extends jni.JObject, $V extends jni.JObject>
+final class $HashMapType<$K extends jni.JObject, $V extends jni.JObject>
     extends jni.JObjType<HashMap<$K, $V>> {
   final jni.JObjType<$K> K;
   final jni.JObjType<$V> V;