+++ jnigen/test/simple_package_test/c_based/dart_bindings/simple_package.dart
@@ -65,7 +65,7 @@ class Color extends jni.JObject {
   }
 }
 
-class $ColorType extends jni.JObjType<Color> {
+final class $ColorType extends jni.JObjType<Color> {
   const $ColorType();
 
   @override
@@ -778,7 +778,7 @@ class Example extends jni.JObject {
   }
 }
 
-class $ExampleType extends jni.JObjType<Example> {
+final class $ExampleType extends jni.JObjType<Example> {
   const $ExampleType();
 
   @override
@@ -863,7 +863,7 @@ class Example_Nested extends jni.JObject {
   }
 }
 
-class $Example_NestedType extends jni.JObjType<Example_Nested> {
+final class $Example_NestedType extends jni.JObjType<Example_Nested> {
   const $Example_NestedType();
 
   @override
@@ -927,7 +927,7 @@ class Example_Nested_NestedTwice extends jni.JObject {
   }
 }
 
-class $Example_Nested_NestedTwiceType
+final class $Example_Nested_NestedTwiceType
     extends jni.JObjType<Example_Nested_NestedTwice> {
   const $Example_Nested_NestedTwiceType();
 
@@ -1003,7 +1003,7 @@ class Example_NonStaticNested extends jni.JObject {
   }
 }
 
-class $Example_NonStaticNestedType
+final class $Example_NonStaticNestedType
     extends jni.JObjType<Example_NonStaticNested> {
   const $Example_NonStaticNestedType();
 
@@ -1248,7 +1248,7 @@ class Exceptions extends jni.JObject {
   }
 }
 
-class $ExceptionsType extends jni.JObjType<Exceptions> {
+final class $ExceptionsType extends jni.JObjType<Exceptions> {
   const $ExceptionsType();
 
   @override
@@ -1503,7 +1503,7 @@ class Fields extends jni.JObject {
   }
 }
 
-class $FieldsType extends jni.JObjType<Fields> {
+final class $FieldsType extends jni.JObjType<Fields> {
   const $FieldsType();
 
   @override
@@ -1593,7 +1593,7 @@ class Fields_Nested extends jni.JObject {
   }
 }
 
-class $Fields_NestedType extends jni.JObjType<Fields_Nested> {
+final class $Fields_NestedType extends jni.JObjType<Fields_Nested> {
   const $Fields_NestedType();
 
   @override
@@ -1657,7 +1657,7 @@ class C2 extends jni.JObject {
   }
 }
 
-class $C2Type extends jni.JObjType<C2> {
+final class $C2Type extends jni.JObjType<C2> {
   const $C2Type();
 
   @override
@@ -1714,7 +1714,7 @@ class Example1 extends jni.JObject {
   }
 }
 
-class $Example1Type extends jni.JObjType<Example1> {
+final class $Example1Type extends jni.JObjType<Example1> {
   const $Example1Type();
 
   @override
@@ -1779,8 +1779,8 @@ class GenericTypeParams<$S extends jni.JObject, $K extends jni.JObject>
   }
 }
 
-class $GenericTypeParamsType<$S extends jni.JObject, $K extends jni.JObject>
-    extends jni.JObjType<GenericTypeParams<$S, $K>> {
+final class $GenericTypeParamsType<$S extends jni.JObject,
+    $K extends jni.JObject> extends jni.JObjType<GenericTypeParams<$S, $K>> {
   final jni.JObjType<$S> S;
   final jni.JObjType<$K> K;
 
@@ -1958,7 +1958,7 @@ class GrandParent<$T extends jni.JObject> extends jni.JObject {
   }
 }
 
-class $GrandParentType<$T extends jni.JObject>
+final class $GrandParentType<$T extends jni.JObject>
     extends jni.JObjType<GrandParent<$T>> {
   final jni.JObjType<$T> T;
 
@@ -2095,8 +2095,8 @@ class GrandParent_Parent<$T extends jni.JObject, $S extends jni.JObject>
   }
 }
 
-class $GrandParent_ParentType<$T extends jni.JObject, $S extends jni.JObject>
-    extends jni.JObjType<GrandParent_Parent<$T, $S>> {
+final class $GrandParent_ParentType<$T extends jni.JObject,
+    $S extends jni.JObject> extends jni.JObjType<GrandParent_Parent<$T, $S>> {
   final jni.JObjType<$T> T;
   final jni.JObjType<$S> S;
 
@@ -2274,7 +2274,7 @@ class GrandParent_Parent_Child<$T extends jni.JObject, $S extends jni.JObject,
   }
 }
 
-class $GrandParent_Parent_ChildType<$T extends jni.JObject,
+final class $GrandParent_Parent_ChildType<$T extends jni.JObject,
         $S extends jni.JObject, $U extends jni.JObject>
     extends jni.JObjType<GrandParent_Parent_Child<$T, $S, $U>> {
   final jni.JObjType<$T> T;
@@ -2380,7 +2380,7 @@ class GrandParent_StaticParent<$S extends jni.JObject> extends jni.JObject {
   }
 }
 
-class $GrandParent_StaticParentType<$S extends jni.JObject>
+final class $GrandParent_StaticParentType<$S extends jni.JObject>
     extends jni.JObjType<GrandParent_StaticParent<$S>> {
   final jni.JObjType<$S> S;
 
@@ -2527,7 +2527,7 @@ class GrandParent_StaticParent_Child<$S extends jni.JObject,
   }
 }
 
-class $GrandParent_StaticParent_ChildType<$S extends jni.JObject,
+final class $GrandParent_StaticParent_ChildType<$S extends jni.JObject,
         $U extends jni.JObject>
     extends jni.JObjType<GrandParent_StaticParent_Child<$S, $U>> {
   final jni.JObjType<$S> S;
@@ -2652,7 +2652,7 @@ class MyMap<$K extends jni.JObject, $V extends jni.JObject>
   }
 }
 
-class $MyMapType<$K extends jni.JObject, $V extends jni.JObject>
+final class $MyMapType<$K extends jni.JObject, $V extends jni.JObject>
     extends jni.JObjType<MyMap<$K, $V>> {
   final jni.JObjType<$K> K;
   final jni.JObjType<$V> V;
@@ -2795,7 +2795,7 @@ class MyMap_MyEntry<$K extends jni.JObject, $V extends jni.JObject>
   }
 }
 
-class $MyMap_MyEntryType<$K extends jni.JObject, $V extends jni.JObject>
+final class $MyMap_MyEntryType<$K extends jni.JObject, $V extends jni.JObject>
     extends jni.JObjType<MyMap_MyEntry<$K, $V>> {
   final jni.JObjType<$K> K;
   final jni.JObjType<$V> V;
@@ -2995,7 +2995,8 @@ class MyStack<$T extends jni.JObject> extends jni.JObject {
   }
 }
 
-class $MyStackType<$T extends jni.JObject> extends jni.JObjType<MyStack<$T>> {
+final class $MyStackType<$T extends jni.JObject>
+    extends jni.JObjType<MyStack<$T>> {
   final jni.JObjType<$T> T;
 
   const $MyStackType(
@@ -3059,7 +3060,7 @@ class StringKeyedMap<$V extends jni.JObject> extends MyMap<jni.JString, $V> {
   }
 }
 
-class $StringKeyedMapType<$V extends jni.JObject>
+final class $StringKeyedMapType<$V extends jni.JObject>
     extends jni.JObjType<StringKeyedMap<$V>> {
   final jni.JObjType<$V> V;
 
@@ -3114,7 +3115,7 @@ class StringMap extends StringKeyedMap<jni.JString> {
   }
 }
 
-class $StringMapType extends jni.JObjType<StringMap> {
+final class $StringMapType extends jni.JObjType<StringMap> {
   const $StringMapType();
 
   @override
@@ -3160,7 +3161,7 @@ class StringStack extends MyStack<jni.JString> {
   }
 }
 
-class $StringStackType extends jni.JObjType<StringStack> {
+final class $StringStackType extends jni.JObjType<StringStack> {
   const $StringStackType();
 
   @override
@@ -3218,7 +3219,7 @@ class StringValuedMap<$K extends jni.JObject> extends MyMap<$K, jni.JString> {
   }
 }
 
-class $StringValuedMapType<$K extends jni.JObject>
+final class $StringValuedMapType<$K extends jni.JObject>
     extends jni.JObjType<StringValuedMap<$K>> {
   final jni.JObjType<$K> K;
 
@@ -3499,7 +3500,7 @@ class _$MyInterfaceImpl<$T extends jni.JObject>
   }
 }
 
-class $MyInterfaceType<$T extends jni.JObject>
+final class $MyInterfaceType<$T extends jni.JObject>
     extends jni.JObjType<MyInterface<$T>> {
   final jni.JObjType<$T> T;
 
@@ -3623,7 +3624,7 @@ class MyInterfaceConsumer extends jni.JObject {
   }
 }
 
-class $MyInterfaceConsumerType extends jni.JObjType<MyInterfaceConsumer> {
+final class $MyInterfaceConsumerType extends jni.JObjType<MyInterfaceConsumer> {
   const $MyInterfaceConsumerType();
 
   @override
@@ -3761,7 +3762,7 @@ class _$MyRunnableImpl implements $MyRunnableImpl {
   }
 }
 
-class $MyRunnableType extends jni.JObjType<MyRunnable> {
+final class $MyRunnableType extends jni.JObjType<MyRunnable> {
   const $MyRunnableType();
 
   @override
@@ -3861,7 +3862,7 @@ class MyRunnableRunner extends jni.JObject {
   }
 }
 
-class $MyRunnableRunnerType extends jni.JObjType<MyRunnableRunner> {
+final class $MyRunnableRunnerType extends jni.JObjType<MyRunnableRunner> {
   const $MyRunnableRunnerType();
 
   @override
@@ -3926,7 +3927,8 @@ class JsonSerializable_Case extends jni.JObject {
   }
 }
 
-class $JsonSerializable_CaseType extends jni.JObjType<JsonSerializable_Case> {
+final class $JsonSerializable_CaseType
+    extends jni.JObjType<JsonSerializable_Case> {
   const $JsonSerializable_CaseType();
 
   @override
@@ -3975,7 +3977,7 @@ class MyDataClass extends jni.JObject {
   }
 }
 
-class $MyDataClassType extends jni.JObjType<MyDataClass> {
+final class $MyDataClassType extends jni.JObjType<MyDataClass> {
   const $MyDataClassType();
 
   @override