+++ jnigen/CHANGELOG.md
@@ -1,4 +1,6 @@
-## 0.6.1-wip
+## 0.7.0-wip
+* **Breaking Change** ([#387](https://github.com/dart-lang/jnigen/issues/387)): Added `JBuffer` and `JByteBuffer` classes as default classes for `java.nio.Buffer` and `java.nio.ByteBuffer` respectively.
+* **Breaking Change**: Made the type classes `final`.
 * Added an explicit cast in generated `<Interface>.implement` code to allow `dart analyze` to pass when `strict-casts` is set.
 
 ## 0.6.0